Add the degm option
This option provides a set of smart quotes that is used in German
typesetting. The same characters as with the French option are used,
however, they are exchanged and set without spacing to the surrounded
text.

I'm sure there is a better way to do this; I have essentially only
copied everything starting with \de and changed it to \degm.